+++ src/main/java/at/hannibal2/skyhanni/config/features/Bingo.java
@@ -1,6 +1,9 @@
 package at.hannibal2.skyhanni.config.features;
 
+import at.hannibal2.skyhanni.config.core.config.Position;
+import at.hannibal2.skyhanni.config.core.config.annotations.ConfigAccordionId;
 import at.hannibal2.skyhanni.config.core.config.annotations.ConfigEditorBoolean;
+import at.hannibal2.skyhanni.config.core.config.annotations.ConfigEditorButton;
 import at.hannibal2.skyhanni.config.core.config.annotations.ConfigOption;
 import com.google.gson.annotations.Expose;
 
@@ -11,4 +14,16 @@ public class Bingo {
             "new area discoveries, and bestiarity upgrades while on bingo.")
     @ConfigEditorBoolean
     public boolean compactChatMessages = true;
+
+    @Expose
+    @ConfigOption(name = "Bingo Card", desc = "Show the bingo card.")
+    @ConfigEditorBoolean
+    @ConfigAccordionId(id = 0)
+    public boolean bingoCard = false;
+
+    @Expose
+    @ConfigOption(name = "Bingo Card Position", desc = "")
+    @ConfigEditorButton(runnableId = "bingoCardPos", buttonText = "Edit")
+    @ConfigAccordionId(id = 0)
+    public Position bingoCardPos = new Position(10, 10, false, true);
 }