+++ src/main/java/at/hannibal2/skyhanni/features/bingo/card/BingoGoals.kt
@@ -1,0 +1,4 @@
+package at.hannibal2.skyhanni.features.bingo.card
+
+
+abstract class BingoGoals(val displayName: String, val description: String, var done: Boolean)