+++ src/main/java/at/hannibal2/skyhanni/features/bingo/nextstep/ProgressionStep.kt
@@ -1,0 +1,3 @@
+package at.hannibal2.skyhanni.features.bingo.nextstep
+
+abstract class ProgressionStep(displayName: String, val amountNeeded: Long, var amountHaving: Long = 0): NextStep(displayName)