+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Features
 + Added highlight for stuff that is missing in the skyblock level guide inventory.
++ Added Bingo Card display.
 
 ## Version 0.15.1 (2023-01-25)